requirements.txt
alembic==1.12.0
attrs==23.1.0
blinker==1.6.2
cffi==1.16.0
click==8.1.7
cryptography==41.0.4
Flask==3.0.0
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1
flask-validate-json==0.0.1
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
Mako==1.2.4
MarkupSafe==2.1.3
marshmallow==3.20.1
packaging==23.2
psycopg2-binary==2.9.9
pycparser==2.21
PyJWT==2.8.0
python-dotenv==1.0.0
referencing==0.30.2
rpds-py==0.10.4
SQLAlchemy==2.0.21
SQLAlchemy-serializer==1.4.1
typing_extensions==4.8.0
webargs==8.3.0
Werkzeug==3.0.0
